Renamed config to config.env
