+++ server/api/jwt/jwt.js
@@ -2,7 +2,7 @@
 const { ExtractJwt } = require("passport-jwt");
 const JwtStrategy = require("passport-jwt").Strategy;
 const passport = require("passport");
-const keys = require("../utils/config");
+const keys = require("../utils/config.env");
 const User = require("../models/user");
 const jwtOptions = {
   jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
